move vibrato code into AkaoSnesTrackLfo.cpp
